WhileAndDoWhileLoopExamples
This project is designed to teach the use of while and do-while loops in C# language. While loop repeats the code block as long as a certain condition is met. The do-while loop is checked after the condition has been executed at least once. This example provides basic examples of how to use both loops.